app用户隐私设置修改接口

+++ api/src/main/java/com/smart/home/controller/pc/UserPrivacyController.java
@@ -54,7 +54,7 @@ public class UserPrivacyController {
 
     @ApiOperation("按主键ID查询用户隐私")
     @GetMapping("/selectById")
-    public APIResponse<UserPrivacy> selectById(Long id) {
+    public APIResponse<UserPrivacy> selectById(Integer id) {
         return APIResponse.OK(userPrivacyService.findById(id));
     }
